% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{add_logo_text}
\alias{add_logo_text}
\title{add_logo_text}
\usage{
add_logo_text(plot, align = 0)
}
\description{
Add UAI CSCN logo to ggplots
}
\examples{
\dontrun{
plot = ggplot2::ggplot(diamonds, ggplot2::aes(price, carat)) + ggplot2::geom_point() +  ggplot2::labs(caption = "@gorkang") + ggplot2::theme_minimal()
add_logo_text(plot)
}
}
