% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_prepare_TASK.R
\name{check_missing_prepare_TASK}
\alias{check_missing_prepare_TASK}
\title{check_missing_prepare_TASK
Downloads all the protocols to CSCN-server
Then checks:
\itemize{
\item for which ones we do not have preparation scripts
\item for which ones we do not have googledoc details in
}
\itemize{
\item https://docs.google.com/spreadsheets/d/1Eo0F4GcmqWZ1cghTpQlA4aHsc8kTABss-HAeimE2IqA/edit#gid=0
\item https://docs.google.com/spreadsheets/d/1LAsyTZ2ZRP_xLiUBkqmawwnKWgy8OCwq4mmWrrc_rpQ/edit#gid=0
}}
\usage{
check_missing_prepare_TASK(
  sync_protocols = FALSE,
  check_trialids = FALSE,
  check_new_task_tabs = FALSE,
  delete_nonexistent = FALSE,
  dont_ask = FALSE,
  show_all_messages = FALSE,
  helper_folder = ".",
  CSCN_server_folder = "..",
  gmail_account = NULL
)
}
\arguments{
\item{sync_protocols}{Download all protocols? TRUE/FALSE}

\item{check_trialids}{Check trialid's are OK? TRUE/FALSE}

\item{check_new_task_tabs}{Check new_tasks Google doc TRUE/FALSE}

\item{delete_nonexistent}{Delete local folders when they do not exist in the server? TRUE/FALSE}

\item{show_all_messages}{TRUE/FALSE}

\item{helper_folder}{By default, the present folder}

\item{CSCN_server_folder}{By default, ../CSCN-server/protocols/}

\item{gmail_account}{Gmail account to connect to google sheets}
}
\description{
check_missing_prepare_TASK
Downloads all the protocols to CSCN-server
Then checks:
\itemize{
\item for which ones we do not have preparation scripts
\item for which ones we do not have googledoc details in
}
\itemize{
\item https://docs.google.com/spreadsheets/d/1Eo0F4GcmqWZ1cghTpQlA4aHsc8kTABss-HAeimE2IqA/edit#gid=0
\item https://docs.google.com/spreadsheets/d/1LAsyTZ2ZRP_xLiUBkqmawwnKWgy8OCwq4mmWrrc_rpQ/edit#gid=0
}
}
