% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status_participants_protocol.R
\name{check_status_participants_protocol}
\alias{check_status_participants_protocol}
\title{check_status_participants_protocol
CHECKS the mysql database and extracts the number of participants completed, discarded and assigned per protocol}
\usage{
check_status_participants_protocol(pid = NULL)
}
\arguments{
\item{pid}{If you want to see a single protocol, enter the protocol id}
}
\description{
check_status_participants_protocol
CHECKS the mysql database and extracts the number of participants completed, discarded and assigned per protocol
}
