% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{check_tasks_source}
\alias{check_tasks_source}
\title{check_tasks_source}
\usage{
check_tasks_source(gmail_account)
}
\arguments{
\item{gmail_account}{Gmail account to download the google sheets}
}
\description{
Checks the different sources of tasks (r package, google sheets with canonical
tasks and google sheets with new tasks), and creates a matrix with all the tasks
}
