% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{CheckLastRun_sync_data_active_protocols}
\alias{CheckLastRun_sync_data_active_protocols}
\title{CheckLastRun_sync_data_active_protocols
Check if DB backup, canonical clean backup and data of active protocols did run today}
\usage{
CheckLastRun_sync_data_active_protocols(abort_if_done = TRUE)
}
\arguments{
\item{abort_if_done}{Abort if everything ran today}
}
\description{
CheckLastRun_sync_data_active_protocols
Check if DB backup, canonical clean backup and data of active protocols did run today
}
\examples{
CheckLastRun_sync_data_active_protocols()
}
