% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{copy_canonical_clean_from_Github_to_server}
\alias{copy_canonical_clean_from_Github_to_server}
\title{copy_canonical_clean_from_Github_to_server}
\usage{
copy_canonical_clean_from_Github_to_server(jsPsych_version = 6, silent = FALSE)
}
\arguments{
\item{jsPsych_version}{6 or 7}

\item{silent}{do not show messages FALSE/TRUE}
}
\description{
copy_canonical_clean_from_Github_to_server
}
