% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jsPsychX_zips.R
\name{create_jsPsychHelpeR_zip}
\alias{create_jsPsychHelpeR_zip}
\title{create_jsPsychHelpeR_zip
Create the jsPsychHelpeR_zip with a template of a full project, needed for the package to run}
\usage{
create_jsPsychHelpeR_zip(add_renv_cache = FALSE)
}
\arguments{
\item{add_renv_cache}{Add the renv cache? Will add couple hundred MB. If TRUE, opening a new project is very fast: TRUE/FALSE}
}
\description{
create_jsPsychHelpeR_zip
Create the jsPsychHelpeR_zip with a template of a full project, needed for the package to run
}
