% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jsPsychX_zips.R
\name{create_jsPsychMonkeys_zip}
\alias{create_jsPsychMonkeys_zip}
\title{create_jsPsychMonkeys_zip
Create the jsPsychMonkeys_zip with a template of a full project, needed for the package to run}
\usage{
create_jsPsychMonkeys_zip(add_renv_cache = FALSE)
}
\arguments{
\item{add_renv_cache}{TRUE/FALSE}
}
\description{
create_jsPsychMonkeys_zip
Create the jsPsychMonkeys_zip with a template of a full project, needed for the package to run
}
