% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protocol_with_missing_in_999.R
\name{create_protocol_with_missing_in_999}
\alias{create_protocol_with_missing_in_999}
\title{create_protocol_with_missing_in_999}
\usage{
create_protocol_with_missing_in_999(
  search_where = "prepare_TASK",
  destination_folder = paste0(tempdir(check = TRUE), "/Missing999")
)
}
\arguments{
\item{search_where}{One of prepare_TASK or js: looks for tasks without results in 999.zip in jsPsychHelpeR/R_tasks/ or in /jsPsychMaker/protocols_DEV/}

\item{destination_folder}{Location of new protocol}
}
\value{
creates a protocol
}
\description{
create_protocol_with_missing_in_999
}
