% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{DELETE_data_server}
\alias{DELETE_data_server}
\title{DELETE_data_server
Delete contents of data folder in the server: /.data/pid}
\usage{
DELETE_data_server(pid = NULL)
}
\arguments{
\item{pid}{Protocol id}
}
\description{
DELETE_data_server
Delete contents of data folder in the server: /.data/pid
}
