% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{download_canonical_clean}
\alias{download_canonical_clean}
\title{download_canonical_clean
Clones from https://github.com/gorkang/jsPsychMaker the canonical_clean folder}
\usage{
download_canonical_clean(
  destination_folder,
  jsPsych_version = 6,
  silent = TRUE
)
}
\arguments{
\item{destination_folder}{Where to download the canonical_clean folder}

\item{jsPsych_version}{jsPsych version (6 or 7)}

\item{silent}{Do not show messages TRUE/FALSE}
}
\description{
download_canonical_clean
Clones from https://github.com/gorkang/jsPsychMaker the canonical_clean folder
}
