% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_check_all_protocols.R
\name{download_check_all_protocols}
\alias{download_check_all_protocols}
\title{SYNC ALL protocols to folder CSCN-server and check}
\usage{
download_check_all_protocols(
  gmail_account,
  ignore_existing = TRUE,
  dont_ask = FALSE,
  show_all_messages = FALSE
)
}
\arguments{
\item{gmail_account}{Gmail account to connect to google sheets}

\item{ignore_existing}{If TRUE, does not overwrite existing files even if they are newer. Good for .data/, Bad for rest}

\item{show_all_messages}{TRUE / FALSE}
}
\description{
Will sync all protocols on server to the LOCAL folder ../CSCN-server
Then, will CHECK:
\itemize{
\item Tasks with no prepare_TASK() script!
\item Tasks NOT in Google Doc
\item Check trialid's are OK
\item Check no missing info in Google doc of NEW tasks
}
}
