% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql_extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{Titextract_tablesle}
\usage{
extract_tables(list_credentials = NULL, serial_parallel = "serial")
}
\arguments{
\item{list_credentials}{list with mysql credentials}

\item{serial_parallel}{serial or parallel}
}
\description{
Titextract_tablesle
}
