% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_github_stars}
\alias{get_github_stars}
\title{get_github_stars}
\usage{
get_github_stars(repo, Renviron)
}
\description{
Gets DF with who starred a Github project
}
\examples{
\dontrun{
DF = get_github_stars(repo = "gorkang/BayesianReasoning",
Renviron = "~/gorkang@gmail.com/RESEARCH/PROYECTOS-Code/jsPsychR/jsPsychAdmin/.Renviron")
}
}
