% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{get_parameters_of_function}
\alias{get_parameters_of_function}
\title{get_parameters_of_function
Prints usage and examples, shows help, and loads usage parameters to Global environment.}
\usage{
get_parameters_of_function(name_function, load_parameters = TRUE)
}
\arguments{
\item{name_function}{Name of the function}

\item{load_parameters}{Load default parameters and their values to the Global Environment}
}
\value{
prints the parameters and examples in the console, shows help, loads the parameters to the Global environment
}
\description{
get_parameters_of_function
Prints usage and examples, shows help, and loads usage parameters to Global environment.
}
\examples{
\dontrun{
get_parameters_of_function(name_function = "jsPsychMaker::create_protocol()", load_parameters = TRUE)
}
}
