% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{simulate_prepare}
\alias{simulate_prepare}
\title{simulate_prepare_template
Simulate and prepare data for a jsPsychMaker folder protocol (or just create the code)}
\usage{
simulate_prepare(
  folder_protocol = NULL,
  n_participants = 100,
  print_watch_run = "print",
  sleep_before_running = 3
)
}
\arguments{
\item{folder_protocol}{jsPsychMaker folder protocol}

\item{n_participants}{Number of participants to simulate}

\item{print_watch_run}{print or run the code}

\item{sleep_before_running}{number of seconds to sleep before running the code}
}
\value{
Will print the jsPsychMonkeys and jsPsychHelpeR commands
}
\description{
simulate_prepare_template
Simulate and prepare data for a jsPsychMaker folder protocol (or just create the code)
}
