% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{rename_jsPsych_canonical}
\alias{rename_jsPsych_canonical}
\title{rename_jsPsych_canonical}
\usage{
rename_jsPsych_canonical(
  canonical_protocol_helper = "~/Downloads/jsPsychR_canonical"
)
}
\arguments{
\item{canonical_protocol_helper}{}
}
\value{
renamed files
}
\description{
rename_jsPsych_canonical
}
