% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-scripts.R
\name{tasks_missing_docs}
\alias{tasks_missing_docs}
\title{tasks_missing_docs
Show list of tasks missing docs/}
\usage{
tasks_missing_docs(
  FOLDER = "~/gorkang@gmail.com/RESEARCH/PROYECTOS-Code/jsPsychR/SHARED-docs/docs"
)
}
\arguments{
\item{FOLDER}{}
}
\description{
tasks_missing_docs
Show list of tasks missing docs/
}
